backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
python-multipart
python-dotenv
motor
pymongo
mtcnn
tensorflow
deepface
tf-keras
opencv-python
scikit-learn
numpy
pandas
reportlab
jinja2
Pillow
